pyproject.toml
[project]
name = "ContactFast"
version = "1.0.0"
description = "FastAPI-powered email submission endpoint for frontend only websites"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi==0.104.1",
    "jinja2>=3.1.0",
    "pg8000>=1.30.0",
    "pydantic[email]==2.5.0",
    "python-dotenv==1.0.0",
    "resend==0.8.0",
    "sqlalchemy>=2.0.35",
    "uvicorn[standard]==0.24.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "black>=23.0.0",
    "mypy>=1.7.0",
]
